Move LegacyExtension to Common - should apply in forgedev too now.
