+++ src/userdev/java/net/minecraftforge/gradle/userdev/UserDevPlugin.java
@@ -12,7 +12,7 @@ import net.minecraftforge.gradle.mcp.MCPRepo;
 import net.minecraftforge.gradle.mcp.tasks.DownloadMCPMappings;
 import net.minecraftforge.gradle.mcp.tasks.GenerateSRG;
 import net.minecraftforge.gradle.userdev.jarjar.JarJarProjectExtension;
-import net.minecraftforge.gradle.userdev.legacy.LegacyExtension;
+import net.minecraftforge.gradle.common.legacy.LegacyExtension;
 import net.minecraftforge.gradle.userdev.tasks.JarJar;
 import net.minecraftforge.gradle.userdev.tasks.RenameJarInPlace;
 import net.minecraftforge.gradle.userdev.util.DeobfuscatingRepo;
@@ -30,15 +30,12 @@ import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.artifacts.DependencySet;
 import org.gradle.api.artifacts.ExternalModuleDependency;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository.MetadataSources;
-import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.SourceDirectorySet;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.provider.Property;
-import org.gradle.api.provider.Provider;
-import org.gradle.api.tasks.JavaExec;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.TaskContainer;
 import org.gradle.api.tasks.TaskProvider;
@@ -304,8 +301,6 @@ public class UserDevPlugin implements Plugin<Project> {
 
             extension.getRuns().forEach(runConfig -> runConfig.token("asset_index", finalAssetIndex));
             Utils.createRunConfigTasks(extension, extractNatives, downloadAssets, createSrgToMcp);
-
-            runRetrogradleFixes(project);
         });
         project.getTasks().withType(JarJar.class).all(jarJar -> {
             logger.info("Creating reobfuscation task for JarJar task: {}", jarJar.getName());
@@ -357,64 +352,4 @@ public class UserDevPlugin implements Plugin<Project> {
         project.getArtifacts().add(JAR_JAR_DEFAULT_CONFIGURATION_NAME, project.getTasks().named(JAR_JAR_TASK_NAME));
     }
 
-    /**
-     * The RetroGradle project aims to port older (below 1.13) versions of Minecraft to the current toolchains.
-     * In the process, some quirks with the older versions of ForgeGradle were discovered that need to be replicated here.
-     *
-     * Each quirk is documented and accounted for in this function.
-     *  - Classpath / Resources;
-     *      FG2 Userdev puts all classes and resources into a single jar file for FML to consume.
-     *      FG3+ puts classes and resources into separate folders, which breaks on older versions.
-     *      We replicate the FG2 behavior by replacing these folders by the jar artifact on the runtime classpath.
-     *  - Dependency AccessTransformers / Coremods;
-     *      FG2 GradleStart exposes a <code>net.minecraftforge.gradle.GradleStart.csvDir</code> property
-     *      that points to a directory containing CSV mappings files. This is used by LegacyDev to remap dependencies' AT modifiers.
-     *      We replicate the FG2 behavior by extracting the mappings to a folder in the build directory
-     *      and setting the property to point to it.
-     *
-     * In other words, it's a containment zone for version-specific hacks.
-     * For issues you think are caused by this function, contact Curle or any other Retrogradle maintainer.
-     */
-    private void runRetrogradleFixes(Project project) {
-        final LegacyExtension config = (LegacyExtension) project.getExtensions().getByName(LegacyExtension.EXTENSION_NAME);
-        // Get the Userdev extension for the run configs
-        final MinecraftExtension minecraft = project.getExtensions().getByType(MinecraftExtension.class);
-        final boolean shouldFixClasspath = config.getFixClasspath().get();
-        final boolean shouldExtractMappings = config.getExtractMappings().get();
-
-        if(shouldFixClasspath) {
-            // create a singleton collection from the jar task's output 
-            final FileCollection jar = project.files(project.getTasks().named("jar"));
-            
-            minecraft.getRuns().stream()
-                // get all RunConfig SourceSets
-                .flatMap(runConfig -> runConfig.getAllSources().stream())
-                .distinct()
-                // replace output directories with the jar artifact on each SourceSet's classpath
-                .forEach(sourceSet -> sourceSet.setRuntimeClasspath(sourceSet.getRuntimeClasspath().minus(sourceSet.getOutput()).plus(jar)));
-        }
-
-        if (shouldExtractMappings) {
-            // Extracts CSV mapping files to a folder in the build directory for further use
-            // by the <code>net.minecraftforge.gradle.GradleStart.csvDir</code> property.
-            final ExtractZip extractMappingsTask = project.getTasks().create("extractMappings", ExtractZip.class, t -> {
-                t.getZip().fileProvider(project.provider(() -> {
-                    // create maven dependency for mappings
-                    String coordinates = MCPRepo.getMappingDep(minecraft.getMappingChannel().get(), minecraft.getMappingVersion().get());
-                    Dependency dep = project.getDependencies().create(coordinates);
-                    // download and cache the mappings
-                    return project.getConfigurations().detachedConfiguration(dep).getSingleFile();
-                }));
-                // mappings are extracted to <root>/build/<taskName> by default
-                t.getOutput().convention(project.getLayout().getBuildDirectory().dir(t.getName()));
-            });
-            final File csvDir = extractMappingsTask.getOutput().get().getAsFile();
-
-            // attach the csvDir location property to each run configuration
-            minecraft.getRuns().configureEach(run -> run.property("net.minecraftforge.gradle.GradleStart.csvDir", csvDir));
-
-            // execute extractMappings before each run task
-            project.getTasks().getByName("prepareRuns").dependsOn(extractMappingsTask);
-        }
-    }
 }
